Buttons connected directly in MaskImageWidget
